backend/requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
flask_cors
flask_migrate
marshmallow-sqlalchemy
PyJWT
psycopg2-binary
Werkzeug
